Fix AMP test suite split (#38)
